vite requires renaming to .jsx
