+++ src/actions/matches-actions.js
@@ -1,5 +1,5 @@
 import base64 from 'base-64';
-import base from '../base';
+import { fetch, update } from '../base';
 import { updateAlert, generateAlertPayload } from './alert-actions';
 
 export const ADD_MATCHES = 'ADD_MATCHES';
@@ -9,6 +9,7 @@ export const UPDATE_MATCH = 'UPDATE_MATCH';
 export const UPDATE_MATCH_SUCCESS = 'UPDATE_MATCH_SUCCESS';
 export const UPDATE_MATCH_FAILURE = 'UPDATE_MATCH_FAILURE';
 export const MATCH_SELECTED = 'MATCH_SELECTED';
+
 
 export const selectMatch = (matchId) => {
   return {
@@ -74,22 +75,23 @@ export const updateMatchFailure = (matchId, err) => {
   };
 };
 
+
 export const fetchMatches = () => {
   return (dispatch) => {
     dispatch(addMatches());
-    return base.fetch('matches', { context: {} })
+    return fetch('matches')
       .then((data) => dispatch(addMatchesSuccess(data)))
       .catch((err) => dispatch(addMatchesFailure(err)));
   };
 };
 
-const fetchAuthCode = () => base.fetch('redemption-codes', { context: {} });
+const fetchAuthCode = () => fetch('redemption-codes');
 const validateRedemptionCode = (userCode, redemptionCode) => base64.encode(userCode.toLowerCase()) === redemptionCode;
 const updateQtyTickets = (data) => {
   if (data.qtyTicketsAvailable > 0) {
     const qty = data.qtyTicketsAvailable - 1
-    return base.update(`matches/${data.id}`, { data: { qtyTicketsAvailable: qty } })
-    .then(() => Promise.resolve(qty))
+    return update(`matches/${data.id}`, { data: { qtyTicketsAvailable: qty } })
+      .then(() => Promise.resolve(qty))
   }
 
   return Promise.resolve()
@@ -101,15 +103,15 @@ export const updateMatchReq = (matchId, payload, userCode) => {
     const defaultError = generateAlertPayload('error', 'Oh no! Something went wrong. Please try again');
 
     dispatch(updateMatch());
-  
+
     return fetchAuthCode()
       .then(redemptionCode => validateRedemptionCode(userCode, redemptionCode))
       .then(validatedCode => {
         if (validatedCode) {
-          return base.fetch(`matches/${matchId}`, { context: {} })
-          .then(updateQtyTickets)
-          .then(qty => dispatch(updateMatchSuccess(matchId, { qtyTicketsAvailable: qty })))
-          .then(() => base.update(`matches/${matchId}`, { data: payload }))
+          return fetch(`matches/${matchId}`, { context: {} })
+            .then(updateQtyTickets)
+            .then(qty => dispatch(updateMatchSuccess(matchId, { qtyTicketsAvailable: qty })))
+            .then(() => update(`matches/${matchId}`, { data: payload }))
         }
         return Promise.reject({ custom: true, msg: 'You did not provide a valid redemption code' });
       })
