+++ src/actions/season-status-actions.js
@@ -1,5 +1,4 @@
-import base from '../base';
-
+import { fetch } from '../base'
 export const ADD_SEASON_STATUS = 'ADD_SEASON_STATUS';
 export const ADD_SEASON_STATUS_SUCCESS = 'ADD_SEASON_STATUS_SUCCESS';
 export const ADD_SEASON_STATUS_FAILURE = 'ADD_SEASON_STATUS_FAILURE';
@@ -36,8 +35,7 @@ export const addSeasonStatusFailure = (data) => {
 export const fetchSeasonStatus = () => {
   return (dispatch) => {
     dispatch(addSeasonStatus());
-    return base
-      .fetch('seasonDelay', { context: {} })
+    return fetch('seasonDelay', { context: {} })
       .then((data) => dispatch(addSeasonStatusSuccess(data)))
       .catch((err) => dispatch(addSeasonStatusFailure(err)));
   };
