--- src/base.js
@@ -1,9 +1,0 @@
-import Rebase from 're-base';
-
-const base = Rebase.createClass({
-  apiKey: 'AIzaSyAGt32afx0uk5uDIJlFxfKnRH0tqgV-jSg',
-  authDomain: 'season-ticket-tracker.firebaseapp.com',
-  databaseURL: 'https://season-ticket-tracker.firebaseio.com',
-});
-
-export default base;
