--- src/setupTests.js
@@ -1,10 +1,0 @@
-const localStorageMock = {
-  getItem: jest.fn((item) => {
-    if (item === 'user') return JSON.stringify({ user: {id: 'taylor' } });
-    return null;
-  }),
-  setItem: jest.fn(),
-  clear: jest.fn()
-};
-
-global.localStorage = localStorageMock;
